+++ src/dbtype/mod.rs
@@ -4,12 +4,12 @@
 
 mod database;
 mod errorresponse;
-mod putdocresponse;
+mod writedocresponse;
 mod viewresult;
 mod viewrow;
 
 pub use self::database::Database;
 pub use self::errorresponse::ErrorResponse;
-pub use self::putdocresponse::PutDocumentResponse;
+pub use self::writedocresponse::{PostToDatabaseResponse, PutDocumentResponse};
 pub use self::viewresult::ViewResult;
 pub use self::viewrow::ViewRow;